Update patch set 2
Patch Set 2:

(4 comments)

> Patch Set 1:
> 
> (4 comments)
> 
> Hi, thanks for contributing
> 
> This test seems a bit more complex than what we can usually manage in Tast, many of its components might adversely affect stability (specifying a golden set of keyboard actions, manual code to get a handle to the keyboard, simultaneously executing graphical apps and keybaord emulation, downloading packages during the test, hard-coding timeouts in shell code). See e.g. crostini.VimCompile, which has failed on ~95% of test runs.
> 
> Can i ask what are you hoping to measure in this test? We might be able to cut out several dependencies by using e.g. a gaming CPU/power benchmark, one that does not require keyboard interaction.
> 
> Related to this, there are several tests for crostini's graphical performance (see src/chromiumos/tast/local/bundles/cros/graphics/crostini_trace_glxgears.go)
> If your intention is to benchmark gaming performance on crostini, it might be best to cc pwang@ or davidriley@ who have expertise in that area.

Our goal is to get a test that comes closer to a real use case. We can have tests like glxgears which are useful synthetic benchmarks, but they are basic and pretty specific to a certain set of operations. In a test with a full game, we get to see performance when all of these graphics computations and UI interactions come together in practice. 

At this point, I agree that it's a little ambitious for Tast, but if we consider this the first baby step to getting Crostini games in the test framework, we could bring in Tast tests for more current games in the future.

Patch-set: 2