arc.Notification: Promote to critical.
According to Tastboard the test is failing frequently on release
builders, but all those failures happen inside the arc.Booted
precondition. This suggests that the test itself is stable but
ARC boot is flaky if it is run in tast.informational-android.
Since there are already many critical ARC tests using the arc.Booted
precondition, promoting this test to critical should not harm CQ
reliability.

BUG=chromium:1029447
TEST=fast_build.sh -T

Change-Id: Idcac882c9063fe95525d3ed208fcaf91dddf126d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1943875
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>