tast-test: Add network.ShillInitScripts test
The Tast test is migrated from network_ShillInitScripts.py, an Autotest.
In the original version, it uses MockkFlimflam instead of shill to run
the following tests: login, loginGuest, loginProfileExists,
loginMultiProfile, and logout.MockFlimflam provides a subset of shill
DBus API so that we can track "dbus-send" invocations sent by the shill
init scripts. These tests are moved to run on real shill in the new Tast
test, because it's possible to observe the expected behaviors without
the mock and that allows for checking the status of the profile stack as
part of these tests. In addition, the function dbusEventMonitor was
added for tracking the "dbus-send" invocations.

BUG=chromium:1026901
TEST=run tast network.ShillInitScripts

Change-Id: Iefad6e6f1aea9ce31db20ce61a16a184bed9fb29