tast-test: Add network.ShillInitLoginScript test
The Tast test is migrated from network_ShillInitScripts.py, an Autotest.
In the original version, it uses MockFlimflam instead of shill to run
the login test. MockFlimflam provides a subset of shill DBus API so that
we can track "dbus-send" invocations sent by the shill init scripts.
These tests are moved to run on real shill in the new Tast test, because
it's possible to observe the expected behaviors without the mock and
that allows for checking the status of the profile stack as part of
these tests. In addition, added the shillscript package that has helper
functions used by the shill init scripts tests and the dbusEventMonitor
function for tracking the "dbus-send" invocations.

BUG=chromium:1026901
TEST=run tast network.ShillInitLoginScripts

Change-Id: Iefad6e6f1aea9ce31db20ce61a16a184bed9fb29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1947713
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>