Update patch set 3
Patch Set 3:

> I'm not sure what the precise outstanding question is here. Can someone reword it more succinctly? Or I'll try to speak my thoughts out loud here.
> 
> > 1) There is no some guarantee how fresh properties are. Properties will be updated only if there is a watcher which is not obvious. It's better to have either immutable properties or updated in the background.
> 
> This seemed odd to me as well, and I was already wondering whether we should, for example, bite the bullet and go all in with watching PropertyChanged signals instead of polling "refresh" methods (e.g., start a watcher in NewManager()?). Then we can remove all these non-obvious caveats of which function refreshes what, and instead just focus on synchronizing the "get"ers with the background property watcher.
> 
> I think that would be inline with what Chrome does (per Matthew) too.
> 
> And I think that would sort of address Oleh's (2)?

Oleh, Matthew, Brian. Thanks for all the comments.
After the discussion, I guess I should review the usages of properties and watchers and see how to redesign these things first.
Personally, I now have some concerns on having the watcher update properties. e.g. spawn a watcher in NewProperties is not very obvious. when the service is not yet ready, we cannot get an initial copy but should be able to spawn watcher. etc. But I'm also not sure if immutable Properties is better.
I guess I'll see how the watcher is used and maybe prepare a doc stating possible change proposal and having it reviewed.

If nothing is missed, I'll abandon this cl tomorrow. Feel free to comment if you have anything to add.

Patch-set: 3