camera.CCAUIExpert: Relax the pattern of metadata file
When two photos are taken in the same seconds, the names would be
"IMG_YYYYMMDD_HHMMSS.jpg" and "IMG_YYYYMMDD_HHMMSS (1).jpg".

So if metadata saving is only enabled for the second photo, the
corresponding metadata json file would be saved without the "(1)" part.

This CL fix the flakiness by relaxing the pattern for the metadata json
file to accept those names.

It's not worth to fix this case in CCA because it's unlikely to happen
in real world, and the renumber logic is in the very late stage of file
saving.

BUG=b:144461721
TEST=Pass camera.CCAUIExpert 10 times in a row.

Change-Id: I6f692c32800ca2ea7e5e8743ed1aad8911f71e0e