tast-test: Add arc.QuarterSizedWindowZooming test
This tast test is for go/quarter-sized-window-zooming feature, in which
the fullscreen window is halved into a quarter window in the Android side
and zoomed to the original size in the Chrome side. In the test app
surface, each row in display pixels is painted in black and white
alternately. And the application of the feature results in alternate color
changes per two rows in pixels.

BUG=b:120237297
TEST=tast run <ip> arc.QuarterSizedWindowZooming
Change-Id: I3edd254e3bd87bdeff04e6557800be42a5c5d6d4