tast-tests: crostini: Update the sanity tests to use parameters
In line with recent changes to the crostini tast tests, we update the
sanity tests to use preconditions (instead of manually duplicating the
setup logic).

We are making this change in advance of promoting the sanity artifact
test to the CQ, so as to avoid renaming it while it runs.

BUG=chromium:1020462
TEST=tast run localhost:33333 crostini.Sanity*

Change-Id: I03cf6c5a2b98bd612e3f5309f9fd7a744bd436db