webrtc: Regard "VideoEncodeAcceleator" suffix as HW encoder
After crrev.com/c/1959234, the name of a used HW encoder is
obtained by .getStats() (e.g. VaapiVideoEncodeAccelerator). The
peerconnection tests has an assumption that the name
"ExternalEncoder" if HW encoder is used.
This removes the assumption; the test judges the HW encoder is
used if the obtained implementation name is "ExternalEncoder"
(for compatibility) or its suffix is "VideoEncodeAccelerator".

BUG=chromium:1031965
TEST=tast run 100.87.84.75 webrtc.RTCPeer* w/wo crrev.com/c/1959234

Change-Id: Iea56f5b4cfb3a07311089fc4d5dd14ec66a6b554