tast-tests: add battery tests for runtime probe
In order to confidently add new probe functions, test cases for runtime
probe should also be implemented.  This CL tests if the battery name in
probe result matches with the battery cros-label which is decoded
by hwid service.  For short-term plan, it is integrated with test_that
cmdline that will pass the filename of cros-labels by |-varsfile|
option.

BUG=b:143126249
TEST=(wilco device) tast -verbose run \
     "-var=autotest.host_info_labels="\
     "[\"hwid_component:battery/<BATTERY NAME>\",\"model:<MODEL NAME>\"]" \
     $ACTIVE_DUT platform.CrosRuntimeProbeBattery

Change-Id: I11b46f0c4c0df499d71090cc29eccddedfcc19db