tast-test: arc.WindowManagerCUJ does not hang on pre-N
This CL calls activity.SetWindowState instead of
ash.SetARCAppWindowState() to prevent "starvation".

ash.SetARCappWindowState() won't receive a "window change" event
when trying to resize a pre-N application since a pop-up dialog
asking for confirmation will appear. And ash.SetARCAppWindowSTate()
will never return.

This CL works-around that by calling activity.SetWindowState() which
is a non-blocking call.

(this CL is part of the needed changes for arc-next)

BUG=b:143908501
TEST=tast run $DUT arc.WindowManagerCUJ

Change-Id: I8bf59b06f6c49a956a19d9a43c6004600d5826d4