tast/webrtc: add black, frozen frame detection logic to DecodePerf
This CL adds logic to detect if a played back video frame is
black or if it's frozen (as compared to the previous one), and
adds those numbers to the perf output. This mimics the code in
peer_connection_perf.go, which will be removed in a later patch.

As a trivia knowledge, frozen frame calculation is based on the
ssim algorithm; both depend on reading the displayed pixels,
for this we use offline canvases to avoid disturbing the main
display thread, and we worked in an 8-downscaled video size.

TEST=tast run -verbose webrtc.DecodePerf.* on kohaku
BUG=chromium:1029548

Change-Id: I3c66619d29ea1cd10f5ab66491aee4bf99aabd9b