Fix ui.WindowResizePerf failures on chromeboxes.
As reported, ui.WindowResizePerf has been failing on chromebox
devices, and that is because presentation callbacks are failed.

I've noticed this happens when no physical displays are connected
to the testing device; which would be the case of testing bot
environment. Indeed, no graphical updates happen on such environment.

With this CL, it checks the existence of physical displays, and
simply returns immediately if no physical displays are connected,
because it can not track meaningful data in such situation.

BUG=chromium:1030113
TEST=tast run with fizz

Change-Id: Ib9d2d7b4ac34c21646809270d62e43988b9949c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1969749
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>