Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5: Code-Review-1
> > 
> > Feedback:
> > 1) I would not run *-risky.txt
> > 2) I would only run *-master.txt
> > 3) I would check all test list in directly. Do not use *.external for this.
> 
> The -1 is for 3). You have a chance to diff test lists across revisions. Use it.

That was the original intent, however as people pointed out (at least we had this comment from nya@ in person), the full test list is about ~9MB in size which is significantly bigger than every other local file in the data directory. 

I'm not against doing it the way you said, as I also agree it's better to have a clear diff of changes instead of binary files, however I'm not sure how acceptable it is to have such large files being pushed to all devices locally instead of being fetched remotely. I'll let the tast team decide on this.

As for risky tests, we can decide what to run depending on what parameterized tast test want to execute (which is the current planned approach). For the full CTS I was under the impression that we'd need to run the risky ones as well. For the CQ tast "quick" run we're most likely going to exclused those however.

Patch-set: 5
Reviewer: Gerrit User 1359131 <1359131@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>