Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > > Patch Set 5: Code-Review-1
> > > > 
> > > > Feedback:
> > > > 1) I would not run *-risky.txt
> > > > 2) I would only run *-master.txt
> > > > 3) I would check all test list in directly. Do not use *.external for this.
> > > 
> > > The -1 is for 3). You have a chance to diff test lists across revisions. Use it.
> > 
> > That was the original intent, however as people pointed out (at least we had this comment from nya@ in person), the full test list is about ~9MB in size which is significantly bigger than every other local file in the data directory. 
> 
> It sounds like nya@ should reconsider. Or maybe you should move the deqp code out of the tast directory into its own repository? There is benefit in being able to run deqp without tast. Most of the world already does so. Tast should not be taken as an excuse to make deqp more obtuse and difficult to run.

I'm not sure I follow, this test is just taking the .apk from the Android CTS and running it automatically against certain devices. It's not making it any easier nor harder to run the deqp itself.

It is simply running 'am start -n com.drawelements.deqp/android.app.NativeActivity  -e 'cmdLine "deqp --deqp-caselist-file=<filename> --deqp-log-filename=<filename>"' 
 
> > I'm not against doing it the way you said, as I also agree it's better to have a clear diff of changes instead of binary files, however I'm not sure how acceptable it is to have such large files being pushed to all devices locally instead of being fetched remotely. I'll let the tast team decide on this.
> 
> These tests run for hours. Google networks can handle 9MB every few hours. I promise.

It was my understanding that the tast files would end up on the test images and that the external ones would be downloaded only during execution of the tests. Again, I don't know if 9MB is *too much* or not but I can totally see the rationale of wanting to keep the local files on the device low so we don't end up filling the root partition with unnecessary files and explode the image size in an uncontrolled manner. I'll let the tast team have a final say on this, I have no particular stake in either approach. 

> > As for risky tests, we can decide what to run depending on what parameterized tast test want to execute (which is the current planned approach). For the full CTS I was under the impression that we'd need to run the risky ones as well. For the CQ tast "quick" run we're most likely going to exclused those however.
> 
> Risky tests destabilize DUTs. Tast does not clean up after broken DUTs. You are creating a problem for every test running after your risky tests. The -master.txt tests at least are somewhat stable. They are stable because mesa and driver vendors agree on keeping them stable. But again, no guarantee. 


If that's the case then I agree that risky tests should be removed, seems like a valid point.

>As for "quick" - I don't understand how you are planing to add many hours of dEQP tests to the CQ? At best you can run a small (carefully selected) number.

Yes, that would be the intention. We'd need to identify a short-enough subset of tests that we can run at a "reasonable" time (10 minutes? whatever the tast team deems appropriate for CQ) to have at least as a simple barrier for bad CLs that prevents them from landing into CQ. As a bare minimum.

Patch-set: 5
Reviewer: Gerrit User 1359131 <1359131@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>