tast-test: Add wlan package for WLAN device info
The package wlan return a pointer to a struct that has the DUT WLAN
device information such as, the name, the product ID, etc.  In addition,
I added code to check for the support of MU-MIMO in wifi_caps.go and
modified wlan_driver.go to make use of the package wlan. The MU-MIMO
capaility is tested by checking if the phy lists the "MU Beamformee"
capability.

BUG=chromium:1032627
TEST=run tast network.WifiCaps and network.WLANDriver

Change-Id: I1dedd7f6645c05621b92d68afc275884b9c58a16