tast-test: Add wlan package for WLAN device info
The package wlan return a pointer to a struct that has the DUT WLAN
device information such as, the name, the product ID, etc.  In addition,
I added code to check for the support of MU-MIMO in wifi_caps.go and
modified wlan_driver.go to make use of the package wlan. The MU-MIMO
capaility is tested by checking if the phy lists the "MU Beamformee"
capability. "MU Beamformee" is reported by checking a binary flag.

BUG=chromium:1032627
TEST=run tast network.WifiCaps and network.WLANDriver

Change-Id: I1dedd7f6645c05621b92d68afc275884b9c58a16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1976075
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>