tast/video: introduce video.Play.*_sw test cases
This CL introduces video.Play.{h264,vp8,vp9}_sw test cases to
play.go.  These tests force not-using the hardware accelerator;
they are targetted to be über-stable and to complement the other
test cases already in the said Play test, e.g.:

- Play.vp8_hw <-- forces and verifies hw decoding, only run where
                  there's hw support.
- Play.vp8    <-- don't force anything, can actually be implemented
                  using sw, hw or even trying hw and falling back.
- Play.vp8_sw <-- forcing sw should work everywhere, including on
                  virtual machines (think chrome CQ).

To do this, a new VerifyNoHWAcceleratorMode and a new Chrome
Precondition are added.

TEST=tast -verbose run ${IP} video.Play.* video.PlayVD.* on kohaku and grunt
BUG=chromium:1035649

Change-Id: Ib43bd942e92bff5d5e943b891ba968efa0b5d794