Add expandPIPviaAltTab to arc.PIP
A PIP window can be expanded via selecting the window in the alt-tab
window cycler. This CL adds the tests that check if this works
properly with both short and long alt-tab. (With short alt-tab,
activation switches immediately while with long alt-tab the UI is
shown and you can select the window to activate.)

BUG=b:142029367
TEST=tast -verbose run my_crbook arc.PIP

Change-Id: I666ecaef270fa3b644daaf692bfdc13adad051bd