camera.CCAUISanity.fake: Promote it to CQ
CCA is undergoing massive refactoring recently, so we would like to
promote the most basic test camera.CCAUISanity.fake to CQ now with 18
consecutive passed versions.

The recent test results can be seen on
https://tastboard.corp.google.com/test/camera.CCAUISanity.fake

It passed in 12769~12786, and the failure in 12787 is a real regression.
That's one reason that why we want to promote it ASAP.

Note that the failure in 12777 on nyan_kitty is a one time flake, the
device rebooted when waiting OOBE.

BUG=b:144747002,b:144747449
TEST=Pass on Nocturne.

Change-Id: Icac23f64098c1c624b17100bc40f06a5ca0cb53b