Create RPC service to check DUT boot mode
Context: We are implementing firmware-testing in Tast.
Many firmware tests rely on the DUT being in dev mode or recovery mode.
This new service lets us verify that the DUT is in the expected mode.
It also creates a new test, CheckBootMode, to exercise this service.

Later, when we are able to switch the DUT into dev/recovery mode, it
will probably make sense to rename this new CheckBootMode test as
SwitchBootMode. But for now, the test's name should accurately reflect
what functionality it exercises.

BUG=chromium:1038115, chromium:1037911
TEST=tast run firmware.CheckBootMode $DUT_IP

Change-Id: Id4fa8bdb56c02269207b5b5342c9b053c27ee27f