Create RPC service to check DUT boot mode
Context: We are implementing firmware-testing in Tast.
Many firmware tests rely on the DUT being in dev mode or recovery mode.
This CL allows us to verify that the DUT is in the expected mode.
It also sets up a new SwitchBootMode test, which we will use when the
mode-switcher has been implemented.

BUG=chromium:1038115, chromium:1037911
TEST=tast run firmware.SwitchBootMode $DUT_IP

Change-Id: Id4fa8bdb56c02269207b5b5342c9b053c27ee27f