graphics.FPS: Add test to check that display FPS is around 60
During bringup, we often find platforms where the numbers of frames
per second is wrong (e.g. 58 fps), due to incorrect panel settings
or computations in the driver.

We load a page that frequently (5ms/200Hz) updates the color of a
tall rectangle (to defeat both PSR and PSR2), and measure
drm_vblank_interval using ftrace.

For now, set relatively tight bounds for FPS (60 +/- 0.2 fps, and
0.1 fps stddev).

BUG=chromium:1031054
TEST=tast run graphics.FPS

Change-Id: I9c4cce56c63ac33d9a5cd5c36231a059861c1b82