Introduce CrashTester wrapper
A future change will require some functions in chromecrash to be
conditionally called before chrome.New(), and then have the results
passed into KillAndGetCrashFiles. To make this easier, I'm making
KillAndGetCrashFiles a member function of a new wrapper. Later, the
wrapper can hold onto the extra information from the early calls so that
it's available during KillAndGetCrashFiles.

As a side benefit, this allows me to reduce the number of places I have
to pass the process type and crash file type around.

BUG=chromium:984807, chromium:1038463
TEST=tast -verbose run --extrauseflags chrome_internal my_crbook
ui.ChromeCrash* (which fails on ui.ChromeCrashLoop due to
https://crbug.com/1038826 and ui.ChromeCrashReporterMetrics.success due
to unable-to-enable-metrics-consent issues)

Change-Id: I46bcaf85cfcba8e9f4c0b846cf71a9ce6cc19e5a