Tast-tests: Add network.WifiScan test.
network.WifiScan gets a WiFi device to assert if it is powered and
sends RequestScan to shill. It is used to detect if the
shill -> wpa_supplicant stack is in sane state. We assume a
Chromebook has only one WiFi device and it should be powered. If not,
something must be terribly wrong (refer to b/146758562).

BUG=chromium:1038551, b:146758562
TEST=Run "tast run $DUT network.WifiScan" on a device with ToT image,
it should be okay. Run the test on a device with 12772.0.0 image, it
should fail.

Change-Id: I3be033fa7d618411ed2e2b4783b838f74452bc48