Update patch set 10
Patch Set 10:

> Hmm, I'm a bit surprised to hear that we have to modify "full-featured" libraries to fit our purpose. Do you know particular features missing in those third party libraries but necessary for us? For example, I thought we can reuse packet parsing code in third-party libraries at least.

I think there is a bit of a misunderstanding here for the "server" part. We don't need a real DHCP server. What we need is a thing that reacts to packets sent by shill and introduces errors, delays, and various other conditions during the DHCP DISCOVER / OFFER / REQUEST exchange and during the lease renewal exchange. No real DHCP server will have these "features".


> I thought we can reuse packet parsing code in third-party libraries at least.

For the protocol part, that is a piece that we could definitely reuse if there is a library that exposes a DHCP packet parser that is easy to use and does not come tied to a full featured control plane. I haven't looked for one. Matthew, did you do a bit of research here ? Go has some builtin DNS code, but unfortunately nothing for DHCP in the standard library.

Patch-set: 10
Reviewer: Gerrit User 1238096 <1238096@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>