crash: Add SenderRetry test.
This is a part of the patches to port logging_CrashSender to Tast.

crash.SenderRetry tests that crash_sender retries uploading a crash
dump if it encounters an error on upload.

crash.SenderRetry corresponds to "sender_send_fails" subtest in
logging_CrashSender. The original test used to not test the retry
behavior, but the new test adds that.

BUG=chromium:978255
TEST=fast_build.sh -T
TEST=tast run kohaku crash.SenderRetry

Change-Id: I47b292f39908ded52d4fe6cd6aa99b9ebf2e8b05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1988048
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>