shill: tolerate error due to race of removed Service
In findMatchingService, tolerate the error due to the race between
Service.GetProperties and service removal as it is possible for a
service to disappear at any time. Currently, not yet any errors are
caused by this. This will be helpful when we start having WiFi tests
creating/removing WiFi in the open environment.

BUG=chromium:1034875
TEST=`tast run $DUT '(!disabled && "name:network.*")'`, tested on
rammus DUT

Change-Id: If7ff4276ee02f8b54c9b3e74a8fcf826c3011952