shill: tolerate some expected errors
In findMatchingService, tolerate the error due to the race between
Service.GetProperties and service removal as it is possible for a
service to disappear at any time. Currently, not yet any errors are
caused by this.

In GetWifiInterface, tolerate the error of invalid shill dbus service,
as the purpose of this function is to handle the case that shill is
not yet ready. This may fix the flakiness in network.IwlwifiPCIRescan
when GetWifiInterface failed on GetProperties at the same time that
shill is restarted by pci rescan. (e.g. failure on astronaut)

BUG=chromium:953497
TEST=`tast run $DUT network.*` on rammus/astronaut DUT

Change-Id: If7ff4276ee02f8b54c9b3e74a8fcf826c3011952