Update patch set 3
Patch Set 3:

> Previously Yen-lin and I thought splitting a big CL into minimal independent chunks helps the review progress.

That can help when there really are independent pieces of a big CL, or if there really are clear logical steps in a refactor. In this case I don't think they were really independent or very clear "separate" steps. But this is all more or less subjective.

> Sure it is an good idea squash CL:1966660, CL:1966661, and CL:1966662 together.

Thanks!

Patch-set: 3