tast-tests: crostini: check disk space in precondition
Since some variants of boards with Crostini enabled do not have
sufficient disk space to fit the full 4 GB VM image that is created by
the precondition, add a runtime check to skip tests using the Crostini
preconditions if insufficient disk space is available.  This is
communicated to the tests via the newly-added Valid field in the PreData
structure; for now, this is only checked in the crostini.Sanity tests in
hopes that they can be promoted to the CQ soon.  Ideally there should be
some way to prevent tast from scheduling the Crostini tests on these
boards, but we don't have that today.

BUG=chromium:1039403
TEST=tast run chromebook crostini.Sanity.artifact

Change-Id: I3fae18f70d051d663fd6763d8c4e8ae539323550