Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > using existing battle-tested library
> 
> I haven't checked that myself. Do we know that for a fact or do we just assume this is the case ? We can't just grab any code off of github and assume it is production ready, right. That's my main concern here, that the tradeoff of removing your own 1kloc code and replace it by a 10k~20kloc lib looks deceitfully attractive, but might actually reveal to be a negative trade. I am not necessarily saying that particular library is bad, I just don't know.

I'm also not convinced that long-term maintainability is that big of an issue for a mock DHCP server used specifically for ten or so tests. As an example, the DHCP test framework that lives in autotest was written in 2012 and has only been touched a handful of times since (and not at all after 2015). In any case, I can take a look at the library you pointed to and see what I can do.

Patch-set: 10