Update patch set 1
Patch Set 1:

(1 comment)

I'm not particularly comfortable with this change, I think marking the tests as pass when they in fact did not run is a problem that may lead to false confidence in releases in the future.

Even if they are marked as skipped this reduces our test coverage on boards that we do actually  support.

Patch-set: 1