tast-test: arc.WindowManagerCUJ add flaky category
The test arc.WindowManagerCUJ tests different parts of the the Android
Window Manager. Some tests have been very stable since they were added,
but others are still flaky for different reasons.

This CL splits the stable vs. flaky tests, so that it is easier to
promote the stable tests to CQ.

TEST=tast run $DUT arc.WindowManagerCUJ arc.WindowManagerCUJ.flaky
BUG=chromium:972897

Change-Id: Id315f75a5f848dd902c3040f5c21c9abd01f7a10