Update patch set 3
Patch Set 3:

(3 comments)

> Patch Set 2:
> 
> Thank you for contributing this proof of concept game test! It is tempting to have a few of these to complement trace based testing.
> 
> Some thoughts:
> 1) Using apt-get install here is problematic the game binary needs to be pinned and ideally better distributed (not a huge problem at < 1MB but generally speaking games are large and change often).

One reason I chose this game is because it has not seen an apt-get version update since October 2018, but it might be a good idea to pin a specific bin. Updates are definitely an issue moving forward with more relevant games, but it seemed better to try gaming out on something without having to worry about updating.

> 2) To pin we probably need to store the versioned binary on Google Storage? We have not thought enough about that yet.
> 3) This binary is probably harmless but generally speaking we are worried about code leaking data of prototype lab devices. IIRC in autotest we have a function to cut network during the binary execution. We may need the same here.

Definitely a valid concern that we share. Do you know of an example of this that I could base a new patch on?

Patch-set: 3