tast-tests: crostini: Report disk avail/used space in precondition
Crostini tends to flake a lot on weaker boards, and more so recently
since crrev.com/c/1949461. as part of investigating this we'd like to
know when the DUT's disk is exhausted.

This CL logs the available and used disk space when we run the
precondition.

BUG=chromium:1039403
TEST=tast run $DUT crostini.Sanity*

Change-Id: I96a724dabf0a71c11a66d298dde7d9692c30ad2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1989648
Commit-Queue: Nic Hollingum <hollingum@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>