Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> As an alternative solution, we can use protocmp (see http://go/go-cmp-proto), but looks like protocmp now only available on non-stable github repo (in ChromeOS we cannot use internal Google source code).
> 
> Unfortunately, if proto messages are not equal, in the diff message we will see XXX_ fields as well.

The protocmp solution seems superior given that it avoids noise when there's actually a difference. It's unfortunate that it's not available yet, but that shouldn't block progress here. My suggestion would be to:

(1) Keep code as is for now
(2) File a tracking bug to move us over to protocmp once it becomes available in our version
(3) Expand the comment to explain the situation so future engineers have a chance to understand this looking at the code (rather than having to dig up code review comment threads), and reference the bug from #2

Patch-set: 2
Reviewer: Gerrit User 1135824 <1135824@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>