Update patch set 3
Patch Set 3:

> Patch Set 2: Verified-1
> 
> (3 comments)
> 
> Real game is not suitable for performance tracking as it will detect any change on the game side. 

This has been a major hoop that I've had to jump through while putting together something like this, and it's why I ultimately used a simple, open-source game with minimal updating. The complexities of using something like a Steam game are prohibitive at this point, but I think this can work, especially if we pin the binary and keep it on Google Storage as Ilja suggested.

> I kind of agree with Nic that this test is putting too much and adversely affect the stability

Pinning the binary version and adding it to Google Storage as Ilja suggested seems like a good solution for this.

> Our current approach is capturing gl-level trace file and replay it through apitrace and other trace program.
> Robert is refactoring the trace to support large scale game trace inside crostini.

As I understand apitrace, it is good for 3D performance analysis, but this test may be a good way to measure gaming power consumption while complementing the game tracing strategy.

Patch-set: 3