dlcservice.go: Use nebraska instead of updateserver
updaterserver was written as a temporary solution until have nebraska
ready to be deployed everywhere for update purposes. This CL removes
updateserver and uses nebraska binary directly on the DUT to perform DLC
installation/updates.

Also get rid of complete.log file. We can just dump the output into the
tast log. The main reason is when the builders fail on this task, on
some instances we can't get the logs anymore. The only thing available
is the stack trace and high level tast log. This way can see the
complete picture easier.

Also writes the nebraska log into tast logs.

BUG=chromium:1034092
TEST=tast --verbose run 100.90.30.183 platform.DLCService

Cq-Depend: chromium:1992677
Change-Id: I37bcb37faf484007f3369f649756d8d0225ee2e5