Update patch set 19
Patch Set 19:

I realized I missed one obvious thing: where's the packet capturer? We expect that for SimpleConnect tests, for diagnostics, and it serves important test functions for several other tests. It also seems like it's an important piece of this design. (I ignored it in my initial CL, because in terms of required Tast features, it isn't much different than the Router.)

To that end: please take a few extra looks at the original Autotest tests, to make sure you're not skipping too many other big things.

Also, while trying not to work this evening, I recalled a thought I had previously: is the Tast Precondition concept useful to us? Particularly, can it save us test time by diverting some of the Router initialization into a test object that can be part of the test setup? IIUC, that means it can be done only once for an entire Tast invocation (several tests). In Autotest, there are a lot of reasons things got really slow, but one of them was that there was a moderate amount of one-time setup that was constantly repeated for every test.

Anyway, that's just a thought, but perhaps something that could be analyzed now, before we try to set this design in stone.

Patch-set: 19
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>