tast-test: fix playstore.InstallApp
When installing an app with playstore, we waited for arc to list the
package before considering it installed. This is incorrect. Arc will
list the package once it is done downloading and started installing.
Instead, we are now waiting for the open button to be enabled. This
happens after the install finishes.

BUG=b:141885513
TEST=tast run <eve> arc.PlayStore

Change-Id: I94dc6776630c3fdd9b513bfd43e1f3eb07cabc0f