Fix ui.WindowResizePerf test combined with ScreenRotationTest on dru
As I debugged and consulted with oshima and afakhry, it turns out
that use of SetTabletModeEnabled(false) to ensuring into clamshell
mode on a tablet device might not be the ideal setup for a tast
test.

This SetTabletModeEnabled makes temporarily into physical clamshell
state, but when it is actually a tablet physically, recomputing
the state might revert some physical state, which causes unexpected
behavior (in my case, causing display rotation to fail).

Instead, it is safer and better to just attach a mouse device,
and to wait for the tablet mode to end.

This means now ash.EnsureTabletModeEnabled might cause further
SetTabletModeEnabled to fail, which is the case of
ui.TabletTransitionPerf. This CL rewrites ui.TabletTransitionPerf too.

BUG=chromium:1040292
TEST=tast run 'ui.*Perf*' on dru
Change-Id: I6431cf5c32f12a5e75d5834ccd5ff3d72696b206