Fix flakiness in broker crash tests
The 'broker' subtests of ui.ChromeCrashLoggedIn and
ui.ChromeCrashNotLoggedIn were flaking because they relied on seeing
"--type=broker" in the command line of the target processes. However,
broker processes are not exec'ed normally; they are just forked and then
use setproctitle() to set some fake command line arguments.
setproctitle() is not 100% reliable, and in cases where it failed, the
test would be unable to find a process to kill.

Replace by looking for messages in the Chrome log which are generated by
the fork().

BUG=chromium:1038463
Fixed: chromium:1038463
TEST=tast -verbose run --extrauseflags chrome_internal  my_crbook ui.ChromeCrashLoggedIn.* ui.ChromeCrashNotLoggedIn.*

Cq-Depend: 2004088
Change-Id: I66655f4bae1a49e9cfe6e24a7987f44eec4af45f